ENH: Use a `requirements.txt` file to list the external dependencies
Use a `requirements.txt` file to list the external dependencies of the
package.

Separating the list of external dependencies from the actual setup
script conforms to Python standard practices, makes things easier to
maintain, and it is a step in the transition to newer PEPs where the
setup system recommeds using a `pyproject.toml` config file.

Documentation:
https://pip.pypa.io/en/stable/reference/requirements-file-format/