ENH: Allow non-identical results in tractogram downsampling
Allow non-identical results in tractogram downsampling: add a flag so
that the random seed value is set to `None` when it is used, and thus,
allow for non-identical results when randomly drawing a subset of
streamlines.

The default behavior is to obtain reproducible (identical) results
across runs after the change introduced in commit fea55a3.