ENH: Add all released Python version above 3.8 until 3.11 to CI
Add all released Python version above 3.8 until 3.11 to CI: WMA is known
to work with Python 3.10, so add this and other versions so that there
can be an indication that the tool can be used with such versions.

Adapt the requirements file to install the appropriate versions for each
Python version.