ENH: Add script availability testing
Add script availability testing: ensure that scripts can be called by
testing the call with the help flag to show the script documentation.

Add the corresponding `__init__.py` files so that the scripts can be
discovered.

Add the necessary package dependencies to discover and run the tests.

Make the GHA workflow actually run the tests: call `pytest` to collect
tests.