BUG: Fix WM diffusion measurement Python script call for Linux
Fix the WM diffusion measurement Python script call in the ORG atlas
application shell script for Linux.

As described in
https://discourse.slicer.org/t/no-csv-file-in-diffusionmessurements-folder-in-wma/28439

the solution for Linux requires providing the Python script with the
path to the 3D Slicer executable together with the `--launch` option and
the path to the tract measurement 3D Slicer CLI executable as the last
positional argument.