STYLE: Prefer using lowercase to name package
Prefer using lowercase to name package in line with other relevant
scientific computing Python packages.

Change it as well in the commented PyPI URL (note that the package was
never been uploaded to PyPI, and thus, changing the package name at this
stage is safe from that point of view).